Split entity_info_template.ts to aws_cloud.ts and k8s.ts
